Feat: 각 Entity 다대일 관계로 수정
Board-Professor 다대일로 수정
그에 따른 DTO 코드 리팩토링 진행
