+++ src/main/java/com/example/gasip/board/dto/BoardReadResponse.java
@@ -1,6 +1,7 @@
-package com.example.gasip.Board.dto;
+package com.example.gasip.board.dto;
 
-import com.example.gasip.Board.model.Board;
+import com.example.gasip.board.model.Board;
+import com.example.gasip.professor.model.Professor;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,8 +11,7 @@ import java.time.LocalDateTime;
 
 @Getter
 @NoArgsConstructor
-public class BoardCreateResponse {
-
+public class BoardReadResponse {
     @NotNull
     private Long postId;
     @NotNull
@@ -23,31 +23,29 @@ public class BoardCreateResponse {
     @NotNull
     private LocalDateTime updateDate;
     @NotNull
-    private Long profId;
+    private Professor professor;
 
     @Builder
-    public BoardCreateResponse(Long postId, String content, Long clickCount, Long likeCount, LocalDateTime regDate, LocalDateTime updateDate, Long profId) {
+    public BoardReadResponse(Long postId, String content, Long clickCount, Long likeCount, LocalDateTime regDate, LocalDateTime updateDate, Professor professor) {
         this.postId = postId;
         this.content = content;
         this.clickCount = clickCount;
         this.likeCount = likeCount;
         this.regDate = regDate;
         this.updateDate = updateDate;
-        this.profId = profId;
+        this.professor = professor;
     }
 
-    //fromEntity메서드 개발
-    public BoardCreateResponse fromEntity(Board board) {
-        return BoardCreateResponse.builder()
+
+    public BoardReadResponse fromEntity(Board board) {
+        return BoardReadResponse.builder()
                 .postId(board.getPostId())
                 .content(board.getContent())
                 .clickCount(board.getClickCount())
                 .likeCount(board.getLikeCount())
                 .regDate(board.getRegDate())
                 .updateDate(board.getUpdateDate())
-                .profId(board.getProfId())
+                .professor(board.getProfessor())
                 .build();
     }
-
-
 }
