+++ src/main/java/com/example/gasip/board/repository/BoardRepository.java
@@ -1,6 +1,6 @@
-package com.example.gasip.Board.repository;
+package com.example.gasip.board.repository;
 
-import com.example.gasip.Board.model.Board;
+import com.example.gasip.board.model.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
